# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Vigilancia - Porterias APP API.",
    "version": "1.0.1",
    "description": "Documentación de Vigilancias - Porterias API"
  },
  "paths": {
...
    "/api/v1/blogs/create": {
      "post": {
        "summary": "Crea un nuevo blog",
        "tags": [
          "Blogs"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBlogDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Blog creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogResponseModel"
                }
              }
            }
          },
          "500": {
            "description": "Error del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "error fetching data"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/blogs/all": {
      "get": {
        "summary": "Obtiene todos los blogs",
        "tags": [
          "Blogs"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Número de página para la paginación"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Número de elementos por página"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string"
            },
            "description": "Campo por el cual ordenar los resultados"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "description": "Orden de clasificación (ascendente o descendente)"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de blogs obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "page": {
                      "type": "string"
                    },
                    "pageSize": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/blogs/update": {
      "put": {
        "summary": "Actualiza un blog existente",
        "tags": [
          "Blogs"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditBlogDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Blog actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditBlogResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/blogs/delete/{blog_id}": {
      "delete": {
        "summary": "Elimina un blog",
        "tags": [
          "Blogs"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "blog_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID del blog a eliminar"
          }
        ],
        "responses": {
          "200": {
            "description": "Blog eliminado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
...
}